iox-eclipse-iceoryx#2157 Add tests and incorporate suggestions from review

+++ iceoryx_posh/include/iceoryx_posh/roudi/introspection_types.hpp
@@ -95,7 +95,7 @@ const capro::ServiceDescription
 struct PortThroughputData
 {
     uint64_t m_publisherPortID{0};
-    uint32_t m_sampleSize{0};
+    uint64_t m_sampleSize{0};
     uint64_t m_chunkSize{0};
     double m_chunksPerMinute{0};
     uint64_t m_lastSendIntervalInNanoseconds{0};
