+++ iceoryx_posh/test/moduletests/test_popo_client_port.cpp
@@ -207,7 +207,7 @@ class ClientPort_test : public Test
 
   public:
     static constexpr uint64_t USER_PAYLOAD_SIZE{32U};
-    static constexpr uint64_t USER_PAYLOAD_ALIGNMENT{8U};
+    static constexpr uint32_t USER_PAYLOAD_ALIGNMENT{8U};
 
     ServerChunkQueueData_t serverChunkQueueData{iox::popo::QueueFullPolicy::DISCARD_OLDEST_DATA,
                                                 iox::popo::VariantQueueTypes::SoFi_MultiProducerSingleConsumer};
